Move server-main to /src
